% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estratofun}
\alias{estratofun}
\title{This function returns the UFAES-Stratum taking stratum as input}
\usage{
estratofun(estrato)
}
\description{
This function returns the UFAES-Stratum taking stratum as input
}

