% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn3fun.R
\name{nn3fun}
\alias{nn3fun}
\title{Returns the nearest neighbour for each element by stratum}
\usage{
nn3fun(id, stratum, values, donor, nvar)
}
\arguments{
\item{id}{Vector with elements identification}

\item{stratum}{Vector with stratum identification}

\item{values}{Vector with variables to compute distance with}

\item{donor}{Class vector for receivers: 0, donors: 1, both: 2}

\item{nvar}{Number of variables in vector values}
}
\value{
list as said
}
\description{
\code{nn3fun} returns a list with the nearest neighbour, order of the donor use, and distance.
}
\details{
Mahalanobis distance is used
}
\examples{
nn3fun
}

